requirements.txt
kafka-python
psycopg2-binary==2.8.6
argparse
textwrap3
faker